+++ doc/changelog.rst
@@ -7,6 +7,8 @@ ChangeLog
 **********************
 
 * Support for exporting keys via :py:func:`gpgliblib.base.GpgKey.export`.
+* Add :py:attr:`gpgliblib.base.GpgKey.keyid` and
+  :py:attr:`gpgliblib.base.GpgKey.long_keyid` convenience attributes.
 
 ****************
 0.1 (2017-01-08)